+++ Loader/AbstractCryptoLoader.php
@@ -1,0 +1,12 @@
+<?php
+namespace Mes\Security\CryptoBundle\Loader;
+
+/**
+ * Class AbstractCryptoLoader
+ *
+ * @package Mes\Security\CryptoBundle\Loader
+ */
+abstract class AbstractCryptoLoader implements CryptoLoaderInterface
+{
+
+}